requirements.txt
nska_deserialize
python-magic
